Remove unnecessary casts
We now all use new enough eclipse versions

+++ src/net/ssehub/kernel_haven/build_model/BuildModelProvider.java
@@ -19,7 +19,7 @@ public class BuildModelProvider extends AbstractProvider<BuildModel> {
 
     @Override
     protected long getTimeout() {
-        return (long) config.getValue(DefaultSettings.BUILD_PROVIDER_TIMEOUT);
+        return config.getValue(DefaultSettings.BUILD_PROVIDER_TIMEOUT);
     }
     
     @Override
